PrePi: Boot UEFI or Android based on cmdline
Now, *IF ENABLED*, the PrePi will try to boot to the appended kernel if the key
combination for recovery mode is *NOT* pressed, even on devices without recovery partitions.
It will also always boot to UEFI if no valid kernel is appended at the end, so that all the current configurations without appending kernel won't get affected.

Signed-off-by: Molly Sophia <mollysophia379@gmail.com>